step of double type

+++ DeathValley/Models/ParabolicChartProcessor.cs
@@ -11,7 +11,7 @@ namespace DeathValley.Models {
 
         public void SetChartPoints(Chart chart) {
             var _points = new List<Chart.Point>();
-            for(int i = chart.LowerLimit; i <= chart.HigherLimit; i+=chart.Step) {
+            for(double i = chart.LowerLimit; i <= chart.HigherLimit; i+=chart.Step) {
                 _points.Add(new Chart.Point { X = i, Y = chart.CoefficientA * i * i + chart.CoefficientB * i + chart.CoefficientC });
             }
             chart.Points = _points;
